Made changes to CreateProfileForm.cs and MacrosForm.cs
CreateProfileForm.cs was changed to InsertNameForm.cs which is now a more generic way of creating forms that are meant for the user to type something into. Title of the form, buttons and txtBox text can now be passed through the constructor's parameters.
Also removed some forms and methods in MacrosForm.cs in order to change the design of creating a new macro